Refatoraçao da main, ajustes nos imports, doc do swagger, inclusao do readme e requirements
