+++ Classes/exportacao.py
@@ -1,34 +1,16 @@
-from file import *
+from Classes.utils import File, RetornoExportacao, RetornoAuxiliar
 
 class FileExportacao(File):
 
     def ordernar_por_nome(linha):
         return linha.lower().split('\t')[1]
 
-    #def LerCsv():
     def get_list_by_csv(self,tipoProduto):
         self.download_csv()
-        """Dicionario = {
-            'Vinhos': _fileNameProd1,
-            'Espumantes': _fileNameProd2,
-            'Uvas_Frescas': _fileNameProd3,
-            'Sucos': _fileNameProd4,
-        }"""
         lista = []
-        #csv_colunas = []
-
-        #for key, val in Dicionario.items():
-            #csv = self.read_csv(val, key != 'Vinhos')
-            #if(key == 'Vinhos'):
 
         csv = self.read_csv(self.fileName,False)
 
-        #if self.fileName.__contains__("Vinhos"):
-        #if tipoProduto == "Vinhos":
-            #csv = self.read_csv(self.fileName,False)
-            #self.csv_colunas = csv[0].split(';')
-        #else:
-            #csv = self.read_csv(self.fileName, True)
         csv_colunas = csv[0].split(';')
         for i in range(1, len(csv)):
             csv_linha = csv[i].split(';')
@@ -37,5 +19,3 @@ class FileExportacao(File):
                 retorno.Dados.append(RetornoAuxiliar(Ano=csv_colunas[j], Valor=csv_linha[j] if csv_linha[j] != '' else 0))
             lista.append(retorno)
         return lista
-
-
