+++ Classes/utils.py
@@ -1,7 +1,6 @@
 import os
 import urllib.request
 from pathlib import Path
-import pandas as pd
 from typing import List
 from pydantic import BaseModel
 
@@ -41,30 +40,16 @@ class File(BaseModel):
     url: str
     list: List
 
-    #retorno = Retorno
-    #retornoAuxiliar = RetornoAuxiliar
-
     def setAttributePath(self):
         self.path = '../csv/' + self.fileName
 
     def setAttributeList(self,list_):
         self.list =  list_
 
-    #def setFilePath(self,url):
-        #self.url = url
-
-    #def uploadFile(self):
-        #self.fileName = pd.read_csv(self.path,sep=';')
-
-    #def getFile(self):
-        #return self.fileName
-
     def download_csv(self):
-        #File.GetCsv('http://vitibrasil.cnpuv.embrapa.br/download/Comercio.csv', self._fileNameProd)
         self.get_csv()
 
     def get_csv(self):
-    # def get_csv(url, name):
         my_file = Path(self.fileName)
         if my_file.is_file():
             return
@@ -73,7 +58,7 @@ class File(BaseModel):
             os.makedirs(_basePath)
         urllib.request.urlretrieve(self.url, _basePath+self.fileName)
 
-    def read_csv(self,fileName, removeHeader=False):
+    def read_csv(self, removeHeader=False):
         self.setAttributePath()
         with open(self.path, 'r', encoding='UTF-8') as file:
             if (removeHeader):
