+++ Readme.md
@@ -1,0 +1,36 @@
+# API - Tech Challenge 
+
+API de consulta dos dados de vinicultura Embrapa.
+
+## Sobre
+
+Este projeto faz parte da primeira entrega do Tech Challenge do curso FIAP MLE - Grupo 46. A API em questão consulta o banco de dados de vinhos, sucos e derivados provenientes do Estado do Rio Grande do Sul. A API é capaz de retornar dados de produção, processamento, comercialização, importação e exportação de bebidas.
+
+## Framework
+O framework escolhido para construçao da API foi o FastAPI, devido a sua alta performance e facilidade na documentaçao usando Swagger. [Documentação do FastAPI](https://fastapi.tiangolo.com/)
+
+## Instalação
+
+Antes de executar a API, deve-se seguir os seguintes passos:
+
+1. Clonar o repositório;
+2. Instalar as dependências com `pip install -r requirements.txt`;
+
+## Uso
+
+Para executar a API, deve-se executar o seguinte comando:
+
+- `uvicorn main:app --reload`
+
+Para acessar a doc do Swagger, digite /docs# na rota.
+
+## Participantes do Projeto
+
+Nome: Barbara Barreto
+Email: barbaraabb19@gmail.com
+
+Nome: Murilo Fischer de Paula Conceicao
+Email: murilofpc@gmail.com
+
+Nome: Sanderlan Martins da Silva
+Email: sanderlanms@gmail.com
